Adding and refactoring tests for Events
